backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
pymongo>=4.6
python-dotenv>=1.0
openai>=1.0
gunicorn>=21.0
gpt4all
certifi>=2024.0.0
google-generativeai>=0.3.0
